backend/pyproject.toml
[tool.poetry]
name = "SRL Chat"
version = "0.1.0"
description = "LLM Agent that conducts an interview with students about their learning strategies."
authors = ["nise <niels.seidel@fernuni-hagen.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.5"
aiofiles = "24.1.0"
aiohttp = "3.9.5"
aiosignal = "1.3.1"
alembic = "1.13.1"
annotated-types = "0.7.0"
anyio = "3.7.1"
attrs = "23.2.0"
blinker = "1.6.3"
certifi = "2024.2.2"
charset-normalizer = "3.3.2"
click = "8.1.7"
colorama = "0.4.6"
discord = "2.3.2"
diskcache = "5.6.3"
distro = "1.9.0"
docker = "7.1.0"
exceptiongroup = "1.2.2"
FLAML = "2.1.2"
Flask = "3.0.3"
Flask-Migrate = "4.0.7"
Flask-SQLAlchemy = "3.1.1"
frozenlist = "1.4.1"
greenlet = "3.0.3"
gunicorn = "22.0.0"
h11 = "0.14.0"
h2 = "4.1.0"
hpack = "4.0.0"
httpcore = "1.0.5"
httpx = "0.27.0"
Hypercorn = "0.17.3"
hyperframe = "6.0.1"
idna = "3.7"
itsdangerous = "2.2.0"
Jinja2 = "3.1.4"
Mako = "1.3.5"
MarkupSafe = "2.1.5"
multidict = "6.0.5"
numpy = "1.26.4"
openai = "1.35.13"
packaging = "24.1"
pandas = "2.2.3"
pgvector = "0.3.3"
priority = "2.0.0"
psycopg2 = "2.9.9"
python-dateutil = "2.9.0.post0"
python-dotenv = "1.0.1"
pytz = "2024.2"
regex = "2024.5.15"
requests = "2.32.3"
setuptools = "70.0.0"
six = "1.16.0"
sniffio = "1.3.1"
SQLAlchemy = "2.0.30"
SQLAlchemy-Utils = "0.41.2"
sqlapagination = "0.0.1"
termcolor = "2.4.0"
tiktoken = "0.7.0"
tqdm = "4.66.4"
tzdata = "2024.1"
urllib3 = "2.2.1"
Werkzeug = "3.0.3"
wsproto = "1.2.0"
yarl = "1.9.4"
ollama = "0.4.7"
langchain-ollama = "0.2.3"
langchain-core = "0.3.40"
pydantic = "2.9.0"
tenacity = "8.1.0"
typing-extensions = "4.12.2"
psycopg2-binary = "^2.9.10"
flask-cors = "^5.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
pytest-timeout = "^2.4.0"

[tool.pytest.ini_options]
testpaths = ["../tests"]
markers = [
    "timeout: per-test timeout in seconds",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
